LayoutTests/editing/execCommand/button-expected.txt

This tests ensures contents inside a button element is properly editable.
You should see "hello, world CyberKit" inside the button and "world" should be bolded.
| <button>
|   "hello, "
|   <b>
|     "<#selection-anchor>world<#selection-focus>"
|   " CyberKit"
